api/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
black==23.12.1
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.105.0
filelock==3.14.0
greenlet==3.0.3
h11==0.14.0
identify==2.5.36
idna==3.7
isort==5.13.2
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.27.0
typing_extensions==4.12.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.25.0
virtualenv==20.26.2
Werkzeug==3.0.3
itsdangerous
httpx